requirements.txt
torch
torchvision
opencv-python
spacy
tensorboard
tensorboardX
pillow
pycocotools
pyyaml
cloudpickle
termcolor
tqdm
matplotlib
scikit-image
click
torch_tb_profiler
accelerate
transformers
open_clip_torch
datasets
diffusers